+++ lib/src/iterable/group.dart
@@ -18,7 +18,7 @@ extension GroupExtension<V> on Iterable<V> {
   Iterable<Group<K, V>> groupBy<K>([K Function(V element)? key]) sync* {
     final iterator = this.iterator;
     if (iterator.moveNext()) {
-      final grouper = key ?? (element) => element as K; // ignore: avoid_as
+      final grouper = key ?? (element) => element as K;
       var group = Group<K, V>(grouper(iterator.current), <V>[iterator.current]);
       while (iterator.moveNext()) {
         final nextKey = grouper(iterator.current);